RavenDB-11310 - Move commands of RachisConsensus to dedicated files
