+++ src/Raven.Server/Rachis/Commands/SwitchToSingleLeaderCommand.cs
@@ -1,0 +1,31 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using Raven.Server.Documents.TransactionMerger.Commands;
+using Raven.Server.ServerWide.Context;
+
+namespace Raven.Server.Rachis.Commands
+{
+    public class SwitchToSingleLeaderCommand : MergedTransactionCommand<ClusterOperationContext, ClusterTransaction>
+    {
+        private readonly RachisConsensus _engine;
+
+        public SwitchToSingleLeaderCommand(RachisConsensus engine)
+        {
+            _engine = engine;
+        }
+
+        protected override long ExecuteCmd(ClusterOperationContext context)
+        {
+            _engine.SwitchToSingleLeader(context);
+            return 1;
+        }
+
+        public override IReplayableCommandDto<ClusterOperationContext, ClusterTransaction, MergedTransactionCommand<ClusterOperationContext, ClusterTransaction>> ToDto(ClusterOperationContext context)
+        {
+            throw new NotImplementedException();
+        }
+    }
+}
